+++ Fox/Public/Reporters/FOXDebugReporter.m
@@ -20,9 +20,9 @@
     return self;
 }
 
-- (void)runnerWillRunWithSeed:(uint32_t)randomSeed
+- (void)runnerWillRunWithSeed:(NSUInteger)randomSeed
 {
-    [self logString:[NSString stringWithFormat:@"Checking with random seed: %u\n", randomSeed]];
+    [self logString:[NSString stringWithFormat:@"Checking with random seed: %lu\n", (unsigned long)randomSeed]];
 }
 
 - (void)runnerWillVerifyTestNumber:(NSUInteger)testNumber withMaximumSize:(NSUInteger)maxSize
